a bug is fixed.
When a grammar component is deleted, a component just below the deleted grammar component ended up inheriting the count of the deleted grammar component.

This was because I put the count of each grammar as state of eachGrammar component. I fixed the bug by keeping both grammar point and its count in object form as a state for App component.